\name{MSDem-package}
\alias{MSDem-package}
\alias{MSDem}
\docType{package}
\title{
\packageTitle{MSDem}
}
\description{
\packageDescription{MSDem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MSDem}
\packageIndices{MSDem}
}
\author{
\packageAuthor{MSDem}

Maintainer: \packageMaintainer{MSDem}
}
%\references{
%}
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
